audit-scanner/demo/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
anthropic>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
